chore: 컴포넌트 디렉토리 구조 변경 (Columns와 Tables 디렉토리 생성)
