+++ client/src/components/Columns/TextColumn.tsx
