+++ client/src/components/Tables/AccountTable.tsx
@@ -1,9 +1,9 @@
 import { ReactElement } from 'react';
 import TableColumn from '../../types/table';
-import ButtonColumn from './ButtonColumn';
-import NumberColumn from './NumberColumn';
-import TextColumn from './TextColumn';
-import Table from '@components/Table/Table';
+import ButtonColumn from '../Columns/ButtonColumn';
+import NumberColumn from '../Columns/NumberColumn';
+import TextColumn from '../Columns/TextColumn';
+import Table from '@components/Tables/Table';
 import { AccountData } from '@type/data';
 
 const accountTableFormat: Array<TableColumn<AccountData>> = [
