+++ client/src/components/Tables/BomTable.tsx
@@ -1,13 +1,13 @@
 import { ReactElement } from 'react';
 import TableColumn from '../../types/table';
-import ButtonColumn from './ButtonColumn';
-import NumberColumn from './NumberColumn';
-import TextColumn from './TextColumn';
-import Table from '@components/Table/Table';
+import ButtonColumn from '../Columns/ButtonColumn';
+import NumberColumn from '../Columns/NumberColumn';
+import TextColumn from '../Columns/TextColumn';
+import Table from '@components/Tables/Table';
 import { BomData, PartData } from '@type/data';
-import InputColumn from './InputColumn';
+import InputColumn from '../Columns/InputColumn';
 import { useLocation } from 'react-router-dom';
-import LinkColumn from './LinkColumn';
+import LinkColumn from '../Columns/LinkColumn';
 
 const bomTableFormat: Array<TableColumn<BomData & PartData>> = [
   { key: 'No.', component: ({ i }) => <NumberColumn>{i + 1}</NumberColumn> },
