+++ client/src/components/Tables/PartIoTable.tsx
@@ -1,13 +1,13 @@
 import { ReactElement } from 'react';
 import TableColumn from '../../types/table';
-import ButtonColumn from './ButtonColumn';
-import NumberColumn from './NumberColumn';
+import ButtonColumn from '../Columns/ButtonColumn';
+import NumberColumn from '../Columns/NumberColumn';
 import TextColumn from './TextColumn';
-import Table from '@components/Table/Table';
-import { ProductIoData } from '@type/data';
-import LinkColumn from './LinkColumn';
+import Table from '@components/Tables/Table';
+import { PartIoData } from '@type/data';
+import LinkColumn from '../Columns/LinkColumn';
 
-const productIoTableFormat: Array<TableColumn<ProductIoData>> = [
+const partIoTableFormat: Array<TableColumn<PartIoData>> = [
   { key: 'No.', component: ({ i }) => <NumberColumn>{i + 1}</NumberColumn> },
   { key: '수정일', component: ({ row }) => <TextColumn>{row.updatedAt.toLocaleDateString()}</TextColumn> },
   { key: '변동량', component: ({ row }) => <TextColumn>{row.quantity}</TextColumn> },
@@ -15,7 +15,11 @@ const productIoTableFormat: Array<TableColumn<ProductIoData>> = [
   { key: '출고', component: ({ row }) => <TextColumn>{row.quantity < 0 ? row.quantity : ''}</TextColumn> },
   {
     key: '비고',
-    component: ({ row }) => <LinkColumn to={'/part'}>{row.productId}</LinkColumn>,
+    component: ({ row }) => <LinkColumn to={'/part'}>{row.partId}</LinkColumn>,
+  },
+  {
+    key: '거래처 이름',
+    component: ({ row }) => <TextColumn>거래처명은 DB 어떤 항목에서 가져와야 하나요.</TextColumn>,
   },
   {
     key: '확정',
@@ -23,7 +27,7 @@ const productIoTableFormat: Array<TableColumn<ProductIoData>> = [
       <ButtonColumn
         color="--color-dark-gray"
         onClick={() => {
-          console.log(`delete ${row.productIoId}`);
+          console.log(`delete ${row.partIoId}`);
         }}
       >
         확정
@@ -32,6 +36,6 @@ const productIoTableFormat: Array<TableColumn<ProductIoData>> = [
   },
 ];
 
-export default function ProductIoTable({ rows }: { rows: ProductIoData[] }): ReactElement {
-  return <Table rows={rows} format={productIoTableFormat} />;
+export default function PartIoTable({ rows }: { rows: PartIoData[] }): ReactElement {
+  return <Table rows={rows} format={partIoTableFormat} />;
 }
