+++ client/src/components/Tables/ProductIoTable.tsx
@@ -1,13 +1,13 @@
 import { ReactElement } from 'react';
 import TableColumn from '../../types/table';
-import ButtonColumn from './ButtonColumn';
-import NumberColumn from './NumberColumn';
-import TextColumn from './TextColumn';
-import Table from '@components/Table/Table';
-import { PartIoData } from '@type/data';
-import LinkColumn from './LinkColumn';
+import ButtonColumn from '../Columns/ButtonColumn';
+import NumberColumn from '../Columns/NumberColumn';
+import TextColumn from '../Columns/TextColumn';
+import Table from '@components/Tables/Table';
+import { ProductIoData } from '@type/data';
+import LinkColumn from '../Columns/LinkColumn';
 
-const partIoTableFormat: Array<TableColumn<PartIoData>> = [
+const productIoTableFormat: Array<TableColumn<ProductIoData>> = [
   { key: 'No.', component: ({ i }) => <NumberColumn>{i + 1}</NumberColumn> },
   { key: '수정일', component: ({ row }) => <TextColumn>{row.updatedAt.toLocaleDateString()}</TextColumn> },
   { key: '변동량', component: ({ row }) => <TextColumn>{row.quantity}</TextColumn> },
@@ -15,11 +15,7 @@ const partIoTableFormat: Array<TableColumn<PartIoData>> = [
   { key: '출고', component: ({ row }) => <TextColumn>{row.quantity < 0 ? row.quantity : ''}</TextColumn> },
   {
     key: '비고',
-    component: ({ row }) => <LinkColumn to={'/part'}>{row.partId}</LinkColumn>,
-  },
-  {
-    key: '거래처 이름',
-    component: ({ row }) => <TextColumn>거래처명은 DB 어떤 항목에서 가져와야 하나요.</TextColumn>,
+    component: ({ row }) => <LinkColumn to={'/part'}>{row.productId}</LinkColumn>,
   },
   {
     key: '확정',
@@ -27,7 +23,7 @@ const partIoTableFormat: Array<TableColumn<PartIoData>> = [
       <ButtonColumn
         color="--color-dark-gray"
         onClick={() => {
-          console.log(`delete ${row.partIoId}`);
+          console.log(`delete ${row.productIoId}`);
         }}
       >
         확정
@@ -36,6 +32,6 @@ const partIoTableFormat: Array<TableColumn<PartIoData>> = [
   },
 ];
 
-export default function PartIoTable({ rows }: { rows: PartIoData[] }): ReactElement {
-  return <Table rows={rows} format={partIoTableFormat} />;
+export default function ProductIoTable({ rows }: { rows: ProductIoData[] }): ReactElement {
+  return <Table rows={rows} format={productIoTableFormat} />;
 }
